requirements.txt
-i https://pypi.org/simple
click==8.1.3
commonmark==0.9.1
pygments==2.12.0
pyyaml==6.0
rich==12.4.4
typer==0.4.1
